remove extranous files
