+++ plugins/fused-ops-and-kernels/src/fms_acceleration_foak/models/llama.py
@@ -17,13 +17,16 @@ from functools import partial
 
 # Third Party
 from transformers.models.llama.modeling_llama import LlamaAttention, LlamaRMSNorm
+from transformers.models.llama.modeling_llama import LlamaMLP
 
 # Local
 from ..kernels.unsloth.cross_entropy_loss import FastCrossEntropyLoss
 from ..kernels.unsloth.rms_layernorm import fast_rms_layernorm
 from ..kernels.unsloth.rope_embedding import fast_rope_embedding
 from .model_patcher import ModelPatcher, ModelPatcherRule, ModelPatcherTrigger
+from .model_patcher import combine_triggers, combine_functions
 from .utils import build_lora_fused_ops, trigger_fused_ops
+from .utils import KEY_QKV, KEY_O, KEY_MLP
 
 # TODO: have a generic version of this rule
 # - do regex on RMSNorm class name
@@ -42,18 +45,51 @@ ModelPatcher.register(
 ModelPatcher.register(
     ModelPatcherRule(
         rule_id="llama-qkvo",
+        trigger=combine_triggers(
+            ModelPatcherTrigger(
+                check=partial(
+                    trigger_fused_ops, attn_cls=LlamaAttention,
+                    submodule_names=["q_proj", "k_proj", "v_proj"],
+                )
+            ),
+            ModelPatcherTrigger(
+                check=partial(
+                    trigger_fused_ops, attn_cls=LlamaAttention,
+                    submodule_names=["o_proj"],
+                )
+            ),
+            logic='OR',
+        ),
+        forward_builder=combine_functions(
+            partial(
+                build_lora_fused_ops,
+                submodule_names=["q_proj", "k_proj", "v_proj"],
+                fused_op=KEY_QKV,
+            ),
+            partial(
+                build_lora_fused_ops,
+                submodule_names=["o_proj"],
+                fused_op=KEY_O,
+            ),
+            logic='APPEND',
+        ),
+        forward_builder_args=["base_type"],
+    )
+)
+
+ModelPatcher.register(
+    ModelPatcherRule(
+        rule_id="llama-mlp",
         trigger=ModelPatcherTrigger(
             check=partial(
-                trigger_fused_ops,
-                attn_cls=LlamaAttention,
-                qkv_module_names=["q_proj", "k_proj", "v_proj"],
-                o_module_name="o_proj",
+                trigger_fused_ops, attn_cls=LlamaMLP,
+                submodule_names=["up_proj", "down_proj", "gate_proj"],
             )
         ),
-        forward_builder=partial(
+        forward_builder= partial(
             build_lora_fused_ops,
-            qkv_module_names=["q_proj", "k_proj", "v_proj"],
-            o_module_name="o_proj",
+            submodule_names=["up_proj", "down_proj", "gate_proj"],
+            fused_op=KEY_MLP,
         ),
         forward_builder_args=["base_type"],
     )
