+++ plugins/fused-ops-and-kernels/src/fms_acceleration_foak/models/mistral.py
@@ -19,21 +19,17 @@ from functools import partial
 from transformers.models.mistral.modeling_mistral import (
     MistralAttention,
     MistralRMSNorm,
+    MistralMLP
 )
 
 # Local
 from ..kernels.unsloth.cross_entropy_loss import FastCrossEntropyLoss
 from ..kernels.unsloth.rms_layernorm import fast_rms_layernorm
-from ..kernels.unsloth.rope_embedding import fast_rope_embedding as _fast_rope_embedding
+from ..kernels.unsloth.rope_embedding import fast_rope_embedding
 from .model_patcher import ModelPatcher, ModelPatcherRule, ModelPatcherTrigger
+from .model_patcher import combine_triggers, combine_functions
 from .utils import build_lora_fused_ops, trigger_fused_ops
-
-
-# NOTE: fast_rope_embedding does not work with position_ids
-# currently they are ignored
-def fast_rope_embedding(Q, K, cos, sin, position_ids=None):
-    return _fast_rope_embedding(Q, K, cos, sin)
-
+from .utils import KEY_QKV, KEY_O, KEY_MLP
 
 # - do regex on RMSNorm class name
 # - check on the tensors required for fast_rms_layernorm
@@ -45,29 +41,59 @@ ModelPatcher.register(
     ),
 )
 
-# - do regex on Attention class name
-# - have a set of qkv / o module names and check on that
 ModelPatcher.register(
     ModelPatcherRule(
         rule_id="mistral-qkvo",
-        trigger=ModelPatcherTrigger(
-            check=partial(
-                trigger_fused_ops,
-                attn_cls=MistralAttention,
-                qkv_module_names=["q_proj", "k_proj", "v_proj"],
-                o_module_name="o_proj",
-            )
+        trigger=combine_triggers(
+            ModelPatcherTrigger(
+                check=partial(
+                    trigger_fused_ops, attn_cls=MistralAttention,
+                    submodule_names=["q_proj", "k_proj", "v_proj"],
+                )
+            ),
+            ModelPatcherTrigger(
+                check=partial(
+                    trigger_fused_ops, attn_cls=MistralAttention,
+                    submodule_names=["o_proj"],
+                )
+            ),
+            logic='OR',
         ),
-        forward_builder=partial(
-            build_lora_fused_ops,
-            qkv_module_names=["q_proj", "k_proj", "v_proj"],
-            o_module_name="o_proj",
+        forward_builder=combine_functions(
+            partial(
+                build_lora_fused_ops,
+                submodule_names=["q_proj", "k_proj", "v_proj"],
+                fused_op=KEY_QKV,
+            ),
+            partial(
+                build_lora_fused_ops,
+                submodule_names=["o_proj"],
+                fused_op=KEY_O,
+            ),
+            logic='APPEND',
         ),
         forward_builder_args=["base_type"],
     )
 )
 
-# - get the module_name and reload on that
+ModelPatcher.register(
+    ModelPatcherRule(
+        rule_id="mistral-mlp",
+        trigger=ModelPatcherTrigger(
+            check=partial(
+                trigger_fused_ops, attn_cls=MistralMLP,
+                submodule_names=["up_proj", "down_proj", "gate_proj"],
+            )
+        ),
+        forward_builder= partial(
+            build_lora_fused_ops,
+            submodule_names=["up_proj", "down_proj", "gate_proj"],
+            fused_op=KEY_MLP,
+        ),
+        forward_builder_args=["base_type"],
+    )
+)
+
 ModelPatcher.register(
     ModelPatcherRule(
         rule_id="mistral-cross-ent",
@@ -79,9 +105,6 @@ ModelPatcher.register(
     )
 )
 
-# - get the module name
-# - check if "apply_rotary_pos_emb" exists
-# - patch
 ModelPatcher.register(
     ModelPatcherRule(
         rule_id="mistral-rope",
