+++ plugins/fused-ops-and-kernels/src/fms_acceleration_foak/models/utils.py
@@ -1,17 +1,3 @@
-# Copyright The FMS HF Tuning Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 # Standard
 from typing import Callable, List, Type
 
@@ -19,14 +5,13 @@ from typing import Callable, List, Type
 import torch
 import os
 
-# Local
-# GPTQ imports
 from ..fused_ops.unsloth_lora.gptq.fast_lora import (
     apply_lora_qkv as fused_op_qkv_gptq,
-    apply_lora_o as fused_op_o_gptq,
+    apply_lora_o_v2 as fused_op_o_gptq,
     apply_lora_mlp as fused_op_mlp_gptq,
 )
 from .model_patcher import ModelPatcherTrigger
+from functools import partial
 
 
 # simple utility function to guess if its lora layer
@@ -35,6 +20,17 @@ def _is_loralayer(module: torch.nn.Module, names: List[str] = None):
         names = ["lora_A", "lora_B", "base_layer"]
     return all(hasattr(module, x) for x in names)
 
+KEY_QKV = 'qkv'
+KEY_O = 'o'
+KEY_MLP = 'mlp'
+
+FUSED_OPS = {
+    'auto_gptq': {
+        KEY_QKV: fused_op_qkv_gptq,
+        KEY_O: fused_op_o_gptq,
+        KEY_MLP: fused_op_mlp_gptq
+    }
+}
 
 # builds a triple of forward functions, that each can be attached
 # on a series of QKV's, where if the first one is called, will call the
@@ -44,70 +40,67 @@ def _is_loralayer(module: torch.nn.Module, names: List[str] = None):
 #       modules are called q_proj, k_proj, and v_proj, respectively.
 # the fused operation can be changed, depending on what the base layer is
 # i.e. gptq or bnb
-def _build_qkv_forwards(
+def _build_fused_forwards(
     attn: torch.nn.Module,
     fused_operation: Callable = fused_op_qkv_gptq,
-    module_names: List[str] = None,
+    submodule_names: List[str] = None,
 ):
-    if module_names is None:
-        module_names = ["q_proj", "k_proj", "v_proj"]
+    # fused opts expected to produce singular or multiple results
+    # module names must be passed in order of what the fused
 
-    Q = K = V = None
-
+    outs = {}
     # the fused operation will be called on first one that passes in the
     # input X.
     # - populates the triple Q, K, V
     # - subsequent calls will be a no-op until ALL Q, K, V get reset to None
     def _fused_op(X):
 
-        nonlocal Q, K, V
-        if Q is None and K is None and V is None:
-            Q, K, V = fused_operation(attn, X)
+        # if all of the outs are not yet populated
+        if all(x not in outs for x in submodule_names):
+            fused_outs = fused_operation(attn, X)
+            try:
+                fused_outs = list(fused_outs) # not sure if this is correct
+            except TypeError:
+                # if fused_outs is not iterable
+                fused_outs = [fused_outs]
+            for n, x in zip(submodule_names, fused_outs):
+                outs[n] = x
 
     # each of these functions
     # - calls the fused op
     # -
-    error_msg = (
-        "QKV fused_op needs to be first reset with sequential calls to each of them"
+
+    def _forward(self, X, name: str):
+        _fused_op(X)
+        assert name in outs, \
+            "Fused_op needs to be first reset with sequential calls to each of them"
+        V = outs[name]
+        del outs[name]
+        return V
+
+    return zip(
+        submodule_names, 
+        [partial(_forward, name=n) for n in submodule_names]
     )
 
-    def _forward_q(self, X):
-        nonlocal Q
-        _fused_op(X)
-        assert Q is not None, error_msg
-        out, Q = Q, None  # unload
-        return out
-
-    def _forward_k(self, X):
-        nonlocal K
-        _fused_op(X)
-        assert K is not None, error_msg
-        out, K = K, None  # unload
-        return out
-
-    def _forward_v(self, X):
-        nonlocal V
-        _fused_op(X)
-        assert V is not None, error_msg
-        out, V = V, None  # unload
-        return out
-
-    return zip(module_names, [_forward_q, _forward_k, _forward_v])
-
-# TODO: add the MLP
 def build_lora_fused_ops(
     attn: torch.nn.Module,
     base_type: str = "auto_gptq",
-    qkv_module_names: List[str] = None,
-    o_module_name: str = "o_proj",
+    submodule_names: List[str] = None,
+    fused_op: str = KEY_QKV,
 ):
-    if qkv_module_names is None:
-        qkv_module_names = ["q_proj", "k_proj", "v_proj"]
+
+    assert len(submodule_names) > 0, \
+        "When building lora fused ops requires more than one submodule."
+
+    if submodule_names is None:
+        submodule_names = ["q_proj", "k_proj", "v_proj"]
+
+    # get the fused op
+    fused_operation = FUSED_OPS[base_type][fused_op]
 
     # handle the QKVs
     if base_type == "auto_gptq":
-        _qkv_fused_op = fused_op_qkv_gptq
-        _o_fused_op = fused_op_o_gptq
 
         # this is required due to this FSDP fix
         # https://github.com/foundation-model-stack/fms-acceleration/pull/15
@@ -129,19 +122,25 @@ def build_lora_fused_ops(
 
             # patch each of the fused ops to view the attributes
             # back into torch.int32
-            # TODO: add the MLP fused op also
-            _qkv_fused_op = patch_forward_to_view_attributes_before_call(
-                _qkv_fused_op, 
+            # - if there are multiple submodules, then we assume that
+            #   'fused_operation' will be called on module that has
+            #   submodules specified in 'submodule_names'.
+            # - otherwise if there is only a single 'submodule_name', then 
+            #   assume that 'fused_operation' called on the submodule specified
+            #   by 'submodule_name' itself
+            if len(submodule_names) > 1:
+                patched_submodule_names = [
+                    n + '.base_layer' for n in submodule_names
+                ]
+            else:
+                # otherwise assume calling on the 'submodule_name' itself
+                # so its just the base layer.
+                patched_submodule_names = 'base_layer'
+
+            fused_operation = patch_forward_to_view_attributes_before_call(
+                fused_operation, 
                 PATCH_FOR_FSDP_TRITON_V2, torch.int32,
-                submodule_names=[
-                    n + '.base_layer' for n in qkv_module_names
-                ],
-                is_method_forward=False,
-            )
-            _o_fused_op = patch_forward_to_view_attributes_before_call(
-                _o_fused_op,
-                PATCH_FOR_FSDP_TRITON_V2, torch.int32,
-                submodule_names='base_layer',
+                submodule_names=patched_submodule_names,
                 is_method_forward=False,
             )
 
@@ -150,28 +149,30 @@ def build_lora_fused_ops(
             f"Cannot build fused ops for base type '{base_type}'."
         )
 
-    trigger_and_forwards = [
-        (ModelPatcherTrigger(check=_is_loralayer, module_name=name), forward)
-        for name, forward in _build_qkv_forwards(
-            attn,
-            fused_operation=_qkv_fused_op,
-            module_names=qkv_module_names,
-        )
-    ]
-
-    # handle the self-attn output
-    _output_module = getattr(attn, o_module_name)
-    if _is_loralayer(_output_module):
-        trigger_and_forwards.append(
-            (
-                ModelPatcherTrigger(check=_is_loralayer, module_name=o_module_name),
-                _o_fused_op,
+    if fused_op == KEY_QKV:
+        return [
+            (ModelPatcherTrigger(check=_is_loralayer, module_name=name), forward)
+            for name, forward in _build_fused_forwards(
+                attn,
+                fused_operation=fused_operation,
+                submodule_names=submodule_names,
             )
-        )
-
-    # return
-    return trigger_and_forwards
-
+        ]
+    elif fused_op == KEY_O:
+        # otherwise its just a single op
+        submodule_names = submodule_names[0]
+        return [
+            (
+                ModelPatcherTrigger(check=_is_loralayer, module_name=submodule_names),
+                fused_operation,
+            )
+        ]
+    elif fused_op == KEY_MLP:
+        # otherwise just return the fused_op that should be attached at the
+        # top MLP level
+        return fused_operation
+    else:
+        raise NotImplementedError(f"Unknown fused op '{fused_op}'")
 
 # trigger if either of the conditions are met
 # 1. qkv all have LoRA adapters for a fused op
@@ -179,16 +180,12 @@ def build_lora_fused_ops(
 def trigger_fused_ops(
     module: torch.nn.Module,
     attn_cls: Type,
-    qkv_module_names: List[str] = None,
-    o_module_name: str = "o_proj",
+    submodule_names: List[str],
 ):
-    if qkv_module_names is None:
-        qkv_module_names = ["q_proj", "k_proj", "v_proj"]
 
-    _o = getattr(module, o_module_name)
-    _qkv = [getattr(module, x) for x in qkv_module_names]
-
-    # trigger on the attention layer
-    return isinstance(module, attn_cls) and (
-        all(_is_loralayer(x) for x in _qkv) or _is_loralayer(_o)
+    # trigger if the module meets the attn class and the submodules 
+    # are all loralayers
+    _mods = [getattr(module, x) for x in submodule_names]
+    return (
+        isinstance(module, attn_cls) and all(_is_loralayer(x) for x in _mods)
     )
